Decrease App content with creation of HomePage
